Refactor test suite and some other changes
These commits are not the greatest, but I want the commits to
follow the same flow as the tutorial, so I am stuck with this. ;)

Tests were split up in subtests, these subtests were later refactored
by breaking out the test logic into "assertCorrectMessage".
DRY it up basically.

t.Helper() is called inside the function to let Go know that
this is a helper method. When this fails the line number reported
will be in the function call rather than inside the test helper, which
is pretty sweet